Return-type hypothesis confirmed: getZoromeTime is int, not u8; WM_START 11/14
Correcting the shared header I got wrong. int rather than u8 eliminated the
spurious byte-mask in both affected functions at once, and unk_17A3C0 then
closed to MATCH. I had applied the u8 on an agent's inference, and a mangled
name cannot carry a return type, so it was never evidence. Five binaries
re-verified green after the correction. Seventh wrong return type on this
project -- a spurious mask or width-narrowing in a diff should now be treated as
an inferred-return-type suspicion before anything else.

IsSingleEntry tested as int with no change, so it stays bool: a neutral result
is not evidence for either type.

The declare-before-the-call lever also proved to apply per value rather than
once per function -- a second argument needed the same treatment as the first.

Two walls characterised rather than forced, including one where the version with
correct values was deliberately kept over the version with the matching branch
type, and one where the agent checked whether the round's new levers applied
before declining to retry.

+++ include/game/bases/d_wm_lib.hpp
@@ -54,7 +54,7 @@ namespace dWmLib {
     void GetStartPointNameFromRouteName(const char *routeName, char *dst);
     void GetEndPointNameFromRouteName(const char *routeName, char *dst);
     bool isKoopaShipAnchor();
-    u8 getZoromeTime();
+    int getZoromeTime();
     u8 getStartPointKinokoHouseKindNum();
     bool isStartPointKinokoHouseStar();
     bool isStartPointKinokoHouse1up();
